Nodes are now obtained, load and save improved
The number of nodes (the order, basically) are now obtained from the spin button, and when the number changes, the grid also dinamically changes. The global variable current_size saves the current number of nodes.

The Graph struct now has an order parameter for the saving and loading functions. loadGraph specifically changes the node number on the spin button depending on the one saved.

#define SIZE 5 is changed to 12 (the maximun allowed) because saving grids with more than 5 nodes caused errors (that define can maybe be erased).