New function for symmetrical matrix (see description)
on_entry_insert_text() now only validates the basic requirements.
This is because of the way gtk works, it can't change 2 entries at the same time, that's why on_entry change() was necessary.
This new function has the diagonal and undirected validation, combined in just one nested for loop. The previous one was in two.